+++ CHANGES.md
@@ -29,8 +29,9 @@
     - `cameraBodyUseRoomOriginCenter`: by default `False`. if `True` Body camera will Use the center of the room as the anchor point instead of moving and rotating avatar, pointing towards the roomspace forward
     - `cameraShoulderUseRoomOriginCenter`: by default `False`. if `True` Shoulder camera will Use the center of the room as the anchor point, instead of the player head. Will however use the player height.
     - `cameraShoulderFollowGaze`: by default `True`. if `False` Shoulder camera will look towards where the player is looking at, instead using the common roomspace forward.
+    - `minimumCameraDistance`: by default `0.2`. Determines the distance the camera will circle around the avatar when passing by, instead of phasing right through them.
 
-    
+
 
 - Brought `FPSPlugin` improvements to `ActionCamera`
 - Brought `DancePlugin` Experiment improvements to `ActionCamera`